% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_store.R
\name{close_corpus_store}
\alias{close_corpus_store}
\title{Close a corpus store connection}
\usage{
close_corpus_store(con)
}
\arguments{
\item{con}{A DBI connection returned by \code{\link[=open_corpus_store]{open_corpus_store()}}.}
}
\value{
Invisibly \code{TRUE}.
}
\description{
Wrapper around \code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}} that accepts connections created by
\code{\link[=open_corpus_store]{open_corpus_store()}}.
}
